feat(home): renaming home page to about
- going with SPA approach
- renaming home page content to about
- moving to having all components in the homepage rather than in homepagecontent file
