+++ src/Components/Skills/Skills.tsx
@@ -1,10 +1,11 @@
 import { FC } from 'react';
-import { Stack } from '@mui/material';
+import { Stack, Typography } from '@mui/material';
 import SkillsToggle from 'Components/Skills/SkilsToggle';
 
 const Skills: FC = () => {
   return (
-    <Stack direction="column" sx={{}}>
+    <Stack direction="column" spacing={4}>
+      <Typography sx={{ fontSize: '2.2rem' }}>Skills</Typography>
       <SkillsToggle />
       {/* <SkillDisplay /> */}
     </Stack>
