+++ src/Pages/Home.tsx
@@ -1,7 +1,7 @@
 import { FC, useRef } from 'react';
 import PageContainerGrid from 'Components/Shared/PageContainerGrid';
 import Navbar from 'Components/Navbar/Navbar';
-import HomePageContent from 'Components/HomePageContent/HomePageContent';
+import About from 'Components/About/About';
 import Footer from 'Components/Footer/Footer';
 
 const Home: FC = () => {
@@ -10,7 +10,8 @@ const Home: FC = () => {
   return (
     <PageContainerGrid>
       <Navbar skillSectionRef={skillSectionRef} />
-      <HomePageContent skillSectionRef={skillSectionRef} />
+      <About skillSectionRef={skillSectionRef} />
+      {/* <Skills /> */}
       <Footer />
     </PageContainerGrid>
   );
